[r2807] Обновление: MediaInfo git-v0.7.98-2-g3436139b; Обновление:ZenLib git-v0.4.36.

+++ src/ExtLib/ZenLib/ZenLib/int128s.cpp
@@ -100,7 +100,7 @@ int128::int128 (const char * sz) throw ()
             ++i;
         };
     };
-    const unsigned int len = strlen(sz);
+    const size_t len = strlen(sz);
     for (; i < len; ++i) {
         unsigned int n = 0;
         if (sz [i] >= '0' && sz [i] <= (('0' + (int) radix) < '9'?('0' + (int) radix):'9')) //if (sz [i] >= '0' && sz [i] <= (('0' + (int) radix) <? '9'))
